# [PERF] product: speed up _create_variant_ids with lots of exclusions