@Aurelienvd 2024-03-22
@Feyensv v15 is still supported and from what I see in master we are still looping through `all_combinations` and call `_is_combination_possible_by_config` on one combination at a time [odoo/addons/product/models/product_template.py](https://github.com/odoo/odoo/blob/af8f31dffffda079b98254dc16ebe20958b85488/addons/product/models/product_template.py#L747)

I can retarget to saas-16.4 as it's the customer version but v15 and v16 databases would lose this optimization.