@Feyensv 2024-03-22
Yeah, I know it's still supported, but with every change come a risk, and unless necessary, I believe that we should avoid making big changes in older versions.

Honestly, having a product with 200k combinations is not a standard use of the product (and won't work well in the configurator/matrix/product page).  I'd even say that fixing it in master is already good enough, but can be okay to fix in the client's version since he gets a timeout.