@Aurelienvd 2024-03-22
@Feyensv The code didn't change much between v15 and v17 so I thought it would be better to merge it in the lowest version. I cannot really tell the customer to wait for v18 upgrade scripts if I retarget to master. But I get your point, if you confirm that you prefer to err on the side of caution then I'll retarget to saas-16.4 🙂